Removed eventID from list
As per discussion on the EPCIS/CBV 2.0 MSWG call on 31st March 2020.
Rationale: If e.g. the eventID field was populated with a UUID by the EPCIS server operated by a third party, it is impossible for the organisation that captured the event in the first place to recreate the pre-hash string e.g. to query for/refer to it.